+++ src/@enums/pages.ts
@@ -1,8 +1,11 @@
 export enum PageUrls {
   HOME = "/",
+  SEARCH = "/search",
   LOGIN = "/accounts/login",
   LOGOUT = "/accounts/logout",
   REGISTER = "/accounts/register",
   ADD_POST = "/compose/post",
   SETTINGS = "/accounts/settings",
+  NOTIFICATIONS = "/accounts/notifications",
+  FOLLOW_REQUESTS = "/accounts/follow-requests",
 }