+++ src/@types/pagination.ts
@@ -1,0 +1,7 @@
+export type Pagination = {
+  limit: number;
+  page: number;
+  totalDocs: number;
+  totalPages: number;
+  nextPage?: number;
+};