+++ src/@types/user.ts
@@ -4,6 +4,10 @@ export type User = {
   fullname: string;
   email: string;
   connectionStatus: number;
+  followersCount: number;
+  followingCount: number;
+  postsCount: number;
+  followRequestsCount?: number;
   createdAt: string;
   updatedAt: string;
 };