+++ src/components/Button/index.tsx
@@ -28,7 +28,7 @@ export default function Button({
       {loading ? (
         // TODO: create component for loading icon
         <svg
-          className="animate-spin h-6 w-6 text-current"
+          className="animate-spin h-5 w-5 text-current"
           xmlns="http://www.w3.org/2000/svg"
           fill="none"
           viewBox="0 0 24 24">