+++ src/components/SafeHydrate/index.tsx
@@ -1,0 +1,17 @@
+import { ReactNode } from "react";
+
+interface Props {
+  children: ReactNode;
+}
+
+const SafeHydrate = ({ children }: Props): JSX.Element => {
+  return (
+    <div suppressHydrationWarning>
+      {typeof document === "undefined" || typeof window === "undefined"
+        ? null
+        : children}
+    </div>
+  );
+};
+
+export default SafeHydrate;