+++ tailwind.config.js
@@ -9,6 +9,9 @@ module.exports = {
   darkMode: "class",
   theme: {
     extend: {
+      fontSize: {
+        tiny: "15px",
+      },
       width: {
         128: "32rem",
       },
@@ -27,7 +30,9 @@ module.exports = {
     },
   },
   variants: {
-    extend: {},
+    extend: {
+      borderWidth: "last",
+    },
   },
   plugins: [require("@tailwindcss/forms")],
 };